apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
pgvector==0.3.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.2
anthropic==0.34.2
openai==1.51.0
redis[hiredis]==5.1.0
arq==0.26.0
apscheduler==3.10.4
stripe==10.11.0
resend==2.3.0
jinja2==3.1.4
posthog==3.7.0
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
